javascipt code for DAO
