--- .env.example
@@ -1,6 +1,0 @@
-GOERLI_RPC_URL='https://alchemy.infura.io/v3/1234567890'
-PRIVATE_KEY='abcdefg'
-
-### Optional
-COINMARKETCAP_API_KEY="YOUR_KEY"
-ETHERSCAN_API_KEY="YOUR_KEY"
