+++ .gitignore
@@ -1,17 +1,95 @@
-node_modules
-.env
-.encryptedKey.json
-coverage
-coverage.json
-typechain
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
-#Hardhat files
-cache
+.encryptedKey.json
+
+# dependencies
+/node_modules
+/.pnp
+.pnp.js
+
+# testing
+/coverage
+
+# next.js
+/.next/
+/out/
+
+# production
+/build
+
+# misc
+.DS_Store
+*.pem
+
+# debug
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+.pnpm-debug.log*
+
+# local env files
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+# vercel
+.vercel
+
+.env 
 artifacts
-proposals.json
-.notes.md
-typechain-types
-gas-report.txt
+cache 
+coverage.json 
+
+front-end/node_modules
+frp
+frpc.ini
+front-end/testMoralis.html
+front-end-typescript/node_modules
+front-end-typescript/nft-marketplace-graph/node_modules
+front-end/nft-marketplace-graph/node_modules
 deployments
-.video.md
-.article.md 
+
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+
+# dependencies
+/node_modules
+/.pnp
+.pnp.js
+
+# testing
+/coverage
+
+# next.js
+/.next/
+/out/
+
+# production
+/build
+
+# misc
+.DS_Store
+*.pem
+
+# debug
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+.pnpm-debug.log*
+
+# local env files
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+# vercel
+.vercel
+
+# typescript
+*.tsbuildinfo
+
+front-end-typescript/.next
+front-end-typescript/node_modules
+front-end-typescript/graph-nft-marketplace/node_modules
+typechain-types
