+++ .prettierrc
@@ -1,8 +1,7 @@
 {
-    "tabWidth": 2,
-    "useTabs": false,
-    "semi": false,
-    "singleQuote": false,
-    "printWidth": 100
-  }
-  
+  "tabWidth": 2,
+  "useTabs": false,
+  "semi": false,
+  "singleQuote": false,
+  "printWidth": 100
+}
