--- .solhint.json
@@ -1,9 +1,0 @@
-{
-  "extends": "solhint:recommended",
-  "rules": {
-    "compiler-version": ["error", "^0.8.0"],
-    "func-visibility": ["warn", { "ignoreConstructors": true }],
-    "prettier/prettier": "warning"
-  },
-  "plugins": ["prettier"]
-}
