+++ DAO.md
@@ -1,0 +1,58 @@
+# DAO voting can be done in 3 major ways:
+
+1. Token Bases Voting-
+
+   - It's the most simplest way, the more token you have the more voting power you'll get.
+
+   Problem;
+
+   - Not a Good way, Becoz Rich people will benefit from this.
+
+2. Skin in the Game Voting-
+
+   - In this method, every vote is mapped to their voter, and If someone try to make bad decision for whatever. His/Her voting power will taken from them.
+
+   Problem;
+
+   - How will you tell if some decision Bad or not.
+
+3. Proof of PersonHood or Participation-
+
+   - This is the most perfect and hardest way, Each voter will have 1 vote, No matter how many wallets they have.
+
+   Problem;
+
+   - How you can tell if someone have multiple wallets or not. Becoz you don't have their identity.
+
+   Solution;
+
+   - Not available yet, But It can be solved with Chainlink somehow.
+
+# Types of Voting by environment:
+
+1. ON-chain voting: (ex- Compound)
+
+   PROS;
+
+   - Transparent, Decentralized, Structure will be easy.
+
+   CONS;
+
+   - You charging 1 million dollar to your community for one change.
+
+   * 10000 people => $100/vote(gas) => $1Million/newChange
+
+<br>
+2. OFF-chain voting: (ex: Snapshot {but he does not use chainlink decentralized orcale})
+
+- It will be decentralized don't worry.
+
+- Voter can sign their vote with their voter signature and instead of throughing On-chain, you can just send it to services like IPFS.
+
+- And then gather the result of that data and then send it to On-chain in one go via orcale like-chainlink to save the Gas.
+
+---
+
+You have the chance to build something like sanpshot but more decentralized.
+
+---
