+++ contracts/GovernanceToken.sol
@@ -1,7 +1,11 @@
 // SPDX-License-Identifier: MIT
-pragma solidity ^0.8.9;
+pragma solidity ^0.8.0;
 
 import "@openzeppelin/contracts/token/ERC20/extensions/ERC20Votes.sol";
+
+// NOTE: Delegate Means => A person authorized to act as representative for another;
+
+// ANSWER: "ERC20Votes" extension keeps a history (checkpoints) of each account’s vote power. Vote power can be delegated either by calling the delegate function directly, or by providing a signature to be used with delegateBySig. Voting power can be queried through the public accessors getVotes and getPastVotes.
 
 contract GovernanceToken is ERC20Votes {
   uint256 public s_maxSupply = 1000000000000000000000000;
