+++ contracts/Main-Governance-Contracts/GovernorContract.sol
@@ -1,47 +1,54 @@
 // SPDX-License-Identifier: MIT
-pragma solidity ^0.8.9;
+pragma solidity ^0.8.4;
 
 import "@openzeppelin/contracts/governance/Governor.sol";
+import "@openzeppelin/contracts/governance/extensions/GovernorSettings.sol";
 import "@openzeppelin/contracts/governance/extensions/GovernorCountingSimple.sol";
 import "@openzeppelin/contracts/governance/extensions/GovernorVotes.sol";
 import "@openzeppelin/contracts/governance/extensions/GovernorVotesQuorumFraction.sol";
 import "@openzeppelin/contracts/governance/extensions/GovernorTimelockControl.sol";
 
+/**
+ * @title This contract will hold all the Voting related functions and stuff like- Voting, Executing, Queing etc.
+ * @dev Contract was created by "Openzeppelin Wizard"
+ */
+
 contract GovernorContract is
   Governor,
+  GovernorSettings,
   GovernorCountingSimple,
   GovernorVotes,
   GovernorVotesQuorumFraction,
   GovernorTimelockControl
 {
-  uint256 public s_votingDelay;
-  uint256 public s_votingPeriod;
-
   constructor(
-    ERC20Votes _token,
+    IVotes _token,
     TimelockController _timelock,
-    uint256 _quorumPercentage,
+    /* TODO : This Parameters are the ones We added to make things customizable instead of hardcode */
+    uint256 _votingDelay,
     uint256 _votingPeriod,
-    uint256 _votingDelay
+    uint256 _quorumPercentage /* NOTE : quorumPercentage => How much percentage of total voters should vote in order to pass the proposal, for ex- 10% out of 100% votes must be available in order to pass the proposal */
   )
     Governor("GovernorContract")
+    GovernorSettings(
+      _votingDelay, /* NOTE 1 block */
+      _votingPeriod, /* NOTE 12 secs for 1 block. So, 50400 blocks = ~ 1 week */
+      0 /* NOTE It's Voting Threshold, which means How much token One should have in order to get voting power */
+    )
     GovernorVotes(_token)
     GovernorVotesQuorumFraction(_quorumPercentage)
     GovernorTimelockControl(_timelock)
-  {
-    s_votingDelay = _votingDelay;
-    s_votingPeriod = _votingPeriod;
-  }
-
-  function votingDelay() public view override returns (uint256) {
-    return s_votingDelay; // 1 = 1 block
-  }
-
-  function votingPeriod() public view override returns (uint256) {
-    return s_votingPeriod; // 45818 = 1 week
-  }
+  {}
 
   // The following functions are overrides required by Solidity.
+
+  function votingDelay() public view override(IGovernor, GovernorSettings) returns (uint256) {
+    return super.votingDelay();
+  }
+
+  function votingPeriod() public view override(IGovernor, GovernorSettings) returns (uint256) {
+    return super.votingPeriod();
+  }
 
   function quorum(uint256 blockNumber)
     public
@@ -50,15 +57,6 @@ contract GovernorContract is
     returns (uint256)
   {
     return super.quorum(blockNumber);
-  }
-
-  function getVotes(address account, uint256 blockNumber)
-    public
-    view
-    override(IGovernor, GovernorVotes)
-    returns (uint256)
-  {
-    return super.getVotes(account, blockNumber);
   }
 
   function state(uint256 proposalId)
@@ -77,6 +75,10 @@ contract GovernorContract is
     string memory description
   ) public override(Governor, IGovernor) returns (uint256) {
     return super.propose(targets, values, calldatas, description);
+  }
+
+  function proposalThreshold() public view override(Governor, GovernorSettings) returns (uint256) {
+    return super.proposalThreshold();
   }
 
   function _execute(
