--- hardhat.config.ts
@@ -1,57 +1,0 @@
-import "@typechain/hardhat"
-import "@nomiclabs/hardhat-waffle"
-import "@nomiclabs/hardhat-etherscan"
-import "@nomiclabs/hardhat-ethers"
-import "hardhat-gas-reporter"
-import "dotenv/config"
-import "solidity-coverage"
-import "hardhat-deploy"
-import "solidity-coverage"
-import { HardhatUserConfig } from "hardhat/config"
-
-const COINMARKETCAP_API_KEY = process.env.COINMARKETCAP_API_KEY || ""
-const GOERLI_RPC_URL =
-  process.env.GOERLI_RPC_URL || "https://eth-goerli.alchemyapi.io/v2/your-api-key"
-const PRIVATE_KEY = process.env.PRIVATE_KEY || "privatKey"
-const ETHERSCAN_API_KEY = process.env.ETHERSCAN_API_KEY || ""
-
-const config: HardhatUserConfig = {
-  defaultNetwork: "hardhat",
-  networks: {
-    hardhat: {
-      chainId: 31337,
-      allowUnlimitedContractSize: true
-    },
-    localhost: {
-      chainId: 31337,
-      allowUnlimitedContractSize: true
-    },
-    goerli: {
-      url: GOERLI_RPC_URL,
-      accounts: [PRIVATE_KEY],
-      chainId: 5,
-    },
-  },
-  solidity: "0.8.9",
-  etherscan: {
-    apiKey: ETHERSCAN_API_KEY,
-  },
-  gasReporter: {
-    enabled: true,
-    currency: "USD",
-    outputFile: "gas-report.txt",
-    noColors: true,
-    // coinmarketcap: COINMARKETCAP_API_KEY,
-  },
-  namedAccounts: {
-    deployer: {
-      default: 0, // here this will by default take the first account as deployer
-      1: 0, // similarly on mainnet it will take the first account as deployer. Note though that depending on how hardhat network are configured, the account 0 on one network can be different than on another
-    },
-  },
-  mocha: {
-    timeout: 200000, // 200 seconds max for running tests
-  },
-}
-
-export default config
