--- helper-functions.ts
@@ -1,19 +1,0 @@
-import { run } from "hardhat"
-
-const verify = async (contractAddress: string, args: any[]) => {
-  console.log("Verifying contract...")
-  try {
-    await run("verify:verify", {
-      address: contractAddress,
-      constructorArguments: args,
-    })
-  } catch (e: any) {
-    if (e.message.toLowerCase().includes("already verified")) {
-      console.log("Already verified!")
-    } else {
-      console.log(e)
-    }
-  }
-}
-
-export default verify
