+++ helper-hardhat-config.js
@@ -1,0 +1,43 @@
+const networkConfig = {
+  default: {
+    name: "hardhat",
+  },
+  31337: {
+    name: "localhost",
+  },
+  5: {
+    name: "goerli",
+  },
+  1: {
+    name: "mainnet",
+  },
+}
+
+const developmentChains = ["hardhat", "localhost"]
+const proposalsFile = "proposals.json"
+
+// Governor Values
+const QUORUM_PERCENTAGE = 4 // Need 4% of voters to pass
+const MIN_DELAY = 3600 // 1 hour - after a vote passes, you have 1 hour before you can enact
+//   VOTING_PERIOD = 45818 // 1 week - how long the vote lasts. This is pretty long even for local tests
+const VOTING_PERIOD = 2 // blocks
+const VOTING_DELAY = 1 // 1 Block - How many blocks till a proposal vote becomes active
+const ADDRESS_ZERO = "0x0000000000000000000000000000000000000000"
+
+const NEW_STORE_VALUE = 77
+const FUNC = "store"
+const PROPOSAL_DESCRIPTION = "Proposal #1 77 in the Box!"
+
+module.exports = {
+  networkConfig,
+  developmentChains,
+  proposalsFile,
+  QUORUM_PERCENTAGE,
+  MIN_DELAY,
+  VOTING_DELAY,
+  VOTING_PERIOD,
+  ADDRESS_ZERO,
+  NEW_STORE_VALUE,
+  FUNC,
+  PROPOSAL_DESCRIPTION,
+}
