+++ package.json
@@ -1,5 +1,7 @@
 {
-  "name": "hardhat-dao-fcc",
+  "name": "dao-template-contract",
+  "author": "Rohit Kumar Suman",
+  "license": "MIT",
   "devDependencies": {
     "@nomiclabs/hardhat-ethers": "npm:hardhat-deploy-ethers@0.3.0-beta.12",
     "@nomiclabs/hardhat-etherscan": "^3.0.0",
@@ -13,6 +15,7 @@
     "chai": "^4.3.6",
     "cross-env": "^7.0.2",
     "dotenv": "^10.0.0",
+    "hardhat-contract-sizer": "^2.4.0",
     "ethereum-waffle": "^3.4.0",
     "ethers": "^5.5.3",
     "fs": "^0.0.1-security",
