--- tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2018",
-    "module": "commonjs",
-    "strict": true,
-    "esModuleInterop": true,
-    "outDir": "dist"
-    // optional:
-    // "suppressImplicitAnyIndexErrors": true
-  },
-  "include": ["./scripts", "./test", "./deploy", "./deploy-helpers"],
-  "files": ["./hardhat.config.ts", "./helper-hardhat-config.ts"]
-}
