+++ utils/move-blocks.js
@@ -1,6 +1,6 @@
-import { network } from "hardhat"
+const { network } = require("hardhat")
 
-export async function moveBlocks(amount: number) {
+const moveBlocks = async function (amount) {
   console.log("Moving blocks...")
   for (let index = 0; index < amount; index++) {
     await network.provider.request({
@@ -9,4 +9,8 @@ export async function moveBlocks(amount: number) {
     })
   }
   console.log(`Moved ${amount} blocks`)
+}
+
+module.exports = {
+  moveBlocks,
 }
