+++ utils/move-time.js
@@ -1,8 +1,12 @@
-import { network } from "hardhat"
+const { network } = require("hardhat")
 
-export async function moveTime(amount: number) {
+const moveTime = async function (amount) {
   console.log("Moving blocks...")
   await network.provider.send("evm_increaseTime", [amount])
 
   console.log(`Moved forward in time ${amount} seconds`)
+}
+
+module.exports = {
+  moveTime,
 }
